+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 4.4.2 (August 6, 2026)
+
 ### Adds
 
 - Adds lazy loading to the `MultiSelect` component when `listOverflow="scroll"`.